+++ src/plugins/oauth2/flow/openIdFlow.ts
@@ -2,6 +2,6 @@ import type * as models from '../../../models/openIdInformation';
 
 export interface OpenIdFlow {
   supportsFlow(flow: string): boolean;
-  getCacheKey(config: models.OpenIdConfiguration): string | false;
+  getCacheKey(config: models.OpenIdConfiguration): string;
   perform(config: models.OpenIdConfiguration, context: models.OpenIdContext): Promise<models.OpenIdInformation | false>;
 }