+++ src/plugins/oauth2/flow/passwordFlow.ts
@@ -10,10 +10,8 @@ class PasswordFlow implements OpenIdFlow {
   }
 
   getCacheKey(config: models.OpenIdConfiguration) {
-    if (assertConfiguration(config, ['tokenEndpoint', 'clientId', 'username', 'password'])) {
-      return `password_${config.variablePrefix}_${config.clientId}_${config.username}_${config.tokenEndpoint}`;
-    }
-    return false;
+    assertConfiguration(config, ['tokenEndpoint', 'clientId', 'username', 'password']);
+    return `password_${config.variablePrefix}_${config.clientId}_${config.username}_${config.tokenEndpoint}`;
   }
 
   async perform(