+++ src/plugins/oauth2/oauth2VariableReplacer.ts
@@ -48,30 +48,29 @@ export async function getOAuth2Response(
   if (openIdFlow) {
     const config = getOpenIdConfiguration(prefix, context.variables);
     const cacheKey = openIdFlow.getCacheKey(config);
-    if (cacheKey) {
-      const tokenExchangeConfig = getOpenIdConfiguration(tokenExchangePrefix, context.variables);
-      let openIdInformation = getSessionOpenIdInformation(cacheKey, tokenExchangePrefix ? tokenExchangeConfig : config);
-      userSessionStore.removeUserSession(cacheKey);
-      if (openIdInformation) {
-        log.trace(`openid refresh token flow used: ${cacheKey}`);
-        openIdInformation = await flows.refreshTokenFlow.perform(openIdInformation, context);
+
+    const tokenExchangeConfig = getOpenIdConfiguration(tokenExchangePrefix, context.variables);
+    let openIdInformation = getSessionOpenIdInformation(cacheKey, tokenExchangePrefix ? tokenExchangeConfig : config);
+    userSessionStore.removeUserSession(cacheKey);
+    if (openIdInformation) {
+      log.trace(`openid refresh token flow used: ${cacheKey}`);
+      openIdInformation = await flows.refreshTokenFlow.perform(openIdInformation, context);
+    }
+    if (!openIdInformation) {
+      log.trace(`openid flow ${flow} used: ${cacheKey}`);
+      openIdInformation = await openIdFlow.perform(config, context);
+      if (openIdInformation && tokenExchangePrefix) {
+        openIdInformation = await flows.TokenExchangeFlow.perform(tokenExchangeConfig, openIdInformation, context);
       }
-      if (!openIdInformation) {
-        log.trace(`openid flow ${flow} used: ${cacheKey}`);
-        openIdInformation = await openIdFlow.perform(config, context);
-        if (openIdInformation && tokenExchangePrefix) {
-          openIdInformation = await flows.TokenExchangeFlow.perform(tokenExchangeConfig, openIdInformation, context);
-        }
+    }
+    if (openIdInformation) {
+      log.trace(`openid flow ${flow} finished`);
+      if (utils.isProcessorContext(context)) {
+        utils.setVariableInContext({ oauth2Session: openIdInformation }, context);
       }
-      if (openIdInformation) {
-        log.trace(`openid flow ${flow} finished`);
-        if (utils.isProcessorContext(context)) {
-          utils.setVariableInContext({ oauth2Session: openIdInformation }, context);
-        }
-        userSessionStore.setUserSession(openIdInformation);
-        keepAlive(cacheKey, context.variables);
-        return openIdInformation;
-      }
+      userSessionStore.setUserSession(openIdInformation);
+      keepAlive(cacheKey, context.variables);
+      return openIdInformation;
     }
   }
   return undefined;