+++ src/store/httpRegion.ts
@@ -1,13 +1,7 @@
 import { Hook, HookCancel } from 'hookpoint';
 
 import * as models from '../models';
-import {
-  addHttpFileRequestClientHooks,
-  isError,
-  toEnvironmentKey,
-  addTestResultToHttpRegion,
-  parseError,
-} from '../utils';
+import { addHttpFileRequestClientHooks, isError, toEnvironmentKey } from '../utils';
 
 export class HttpRegion implements models.HttpRegion {
   request?: models.Request<string, models.RequestBody> | undefined;
@@ -63,7 +57,7 @@ export class HttpRegion implements models.HttpRegion {
     isMainContext?: boolean
   ): Promise<boolean> {
     delete this.response;
-    delete this.testResults;
+    this.testResults = [];
 
     if (context.httpRegion) {
       this.registerRegionDependent(context.httpRegion);
@@ -98,17 +92,12 @@ export class HttpRegion implements models.HttpRegion {
       return result !== HookCancel && result.every(obj => !!obj);
     } catch (err) {
       if (isError(err)) {
-        addTestResultToHttpRegion(this, {
-          message: err.message,
-          error: parseError(err),
-          status: models.TestResultStatus.ERROR,
-        });
         if (!err.handled) {
           await context.logResponse?.(this.response, this);
           err.handled = true;
         }
       }
-      throw err;
+      return false;
     }
   }
   private dependentsPerEnv: Array<models.HttpRegion> = [];